requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
openai==1.66.5
numpy==1.26.4
pickle-mixin
faiss-cpu==1.10.0
pypdf==5.2.0
langchain==0.3.20
sentence-transformers==3.4.1
langchain-openai==0.3.9
gunicorn